possibly working

+++ backend/index.js
@@ -2,8 +2,7 @@ import express from 'express';
 import passport from 'passport';
 import jwt from 'jsonwebtoken';
 import session from 'express-session';
-//import KeycloakBearerStrategy from 'passport-keycloak-bearer'
-import  KeycloakBearerStrategy  from 'passport-keycloak-bearer';
+import KeycloakBearerStrategy  from 'passport-keycloak-bearer';
 import passportJWT from 'passport-jwt'
 import DB from './db.js'
 
@@ -23,42 +22,43 @@ app.use(express.json());
 
 /** global instance of our database */
 let db = new DB();
+app.use(session({
+   secret: 'my-secret-key',
+   resave: false,
+   saveUninitialized: false
+}));
 
+app.use(passport.initialize());
+app.use(passport.session());
 
-  app.use(session({
-    secret: 'my-secret-key',
-    resave: false,
-    saveUninitialized: false
-  }));
-  app.use(passport.initialize());
-  app.use(passport.session());
+passport.serializeUser((user, done) => done(null, user))
+passport.deserializeUser((user, done) => done(null, user))
 
 passport.use(new KeycloakBearerStrategy({
    realm: 'webentwicklung',
    url: 'https://jupiter.fh-swf.de/keycloak'
 }, (token, done) => {
    try { 
-      console.log(token)
+      //const tokenStr = JSON.stringify(token);
+      //const decoded = jwt.verify(tokenStr, publicKey, { algorithms: ['RS256'] });
 
-      const decoded = jwt.verify(token, publicKey, { algorithms: ['RS256'] });
-      return done(null, decoded);
+      db.getUserByName(token.preferred_username)
+        .then((rslt) => {return done(null, rslt);})
+        .catch((err) => {return done(null, false);});
+      
     } catch (error) {
-      console.log("bla");
       console.error(error);
       return done(null, false);
     }
 }));
 
 
-app.get('/protected', passport.authenticate('keycloak', { session: true }), (req, res) => {
-   console.log("qqq")
+const authmw =passport.authenticate('keycloak', { session: true })
+
+app.get('/protected', authmw, (req, res) => {
+   console.log(req.user)
    res.send('Hello from protected area!');
 });
-
-
-
-
-
 
 
 /** Initialize database connection */
@@ -67,39 +67,58 @@ async function initDB() {
     console.log("Connected to database");
 }
 
+async function checkTDAccess(todoID,userID){
+    let todo = await db.queryById(todoID);
+    if(todo.user_id === userID){
+        return true;
+    }
+    return false;
 
-
-
+}
 
 // implement API routes
 
 /** Return all todos. 
  *  Be aware that the db methods return promises, so we need to use either `await` or `then` here! 
  */
-app.get('/todos', async (req, res) => {
+app.get('/todos', authmw, async (req, res) => {
     let todos = await db.queryAll();
     res.send(todos);
 });
 
 /** Return todo by given id. 
  */
-app.get('/todos/:id', async (req, res) => {
+app.get('/todos/:id', authmw ,async (req, res) => {
+    
+    if(! (await checkTDAccess(req.params.id,req.user.id))){
+        res.sendStatus(401);
+        return;
+    }
     let todos = await db.queryById(req.params.id);
-    console.log(todos)
     res.send(todos);
 });
 
-app.put('/todo/:id' , async(req,res) =>{
+
+
+app.put('/todo/:id' ,authmw, async(req,res) =>{
+    if(! (await checkTDAccess(req.params.id,req.user.id))){
+        res.sendStatus(401);
+        return;
+    }
     let ret = await db.update(req.params.id, req.body);
     res.send(ret);
 });
 
-app.post('/todo', async(req,res) =>{
+app.post('/todo', authmw, async(req,res) =>{ 
     let ret = await db.insert(req.body);
     res.send(ret);
 });
 
-app.delete('/todo/:id', async(req,res) =>{
+app.delete('/todo/:id', authmw,async(req,res) =>{
+    if(! (await checkTDAccess(req.params.id,req.user.id))){
+        res.sendStatus(401);
+        return;
+    }
     let ret = await db.delete(req.params.id);
     res.send(ret);
 })
@@ -108,8 +127,6 @@ app.delete('/todo/:id', async(req,res) =>{
 app.get('/', async (req, res) => {
     res.send("nix los");
 });
-
-
